+++ common/src/main/java/com/human/common/gameplay/level/patrol/PatrolSpawnerTicker.java
@@ -91,7 +91,7 @@ public class PatrolSpawnerTicker {
                 .toList();
 
             if (players.isEmpty()) {
-                return null;
+                return List.of();
             }
 
             var randomIndex = level.random.nextInt(players.size());